ticket-integration refactor for ArchUnit
