+++ digiwf-integrations/digiwf-ticket-integration/digiwf-ticket-integration-core/src/test/java/de/muenchen/oss/digiwf/ticket/integration/application/WriteArticleUseCaseTest.java
@@ -2,6 +2,7 @@ package de.muenchen.oss.digiwf.ticket.integration.application;
 
 import de.muenchen.oss.digiwf.ticket.integration.application.port.out.LoadFileOutPort;
 import de.muenchen.oss.digiwf.ticket.integration.application.port.out.TicketOutPort;
+import de.muenchen.oss.digiwf.ticket.integration.application.usecase.WriteArticleUseCase;
 import de.muenchen.oss.digiwf.ticket.integration.domain.model.Article;
 import de.muenchen.oss.digiwf.ticket.integration.domain.model.FileContent;
 import de.muenchen.oss.digiwf.ticket.integration.domain.model.TicketStatus;
