+++ digiwf-integrations/digiwf-ticket-integration/digiwf-ticket-integration-starter/src/main/java/de/muenchen/oss/digiwf/ticket/integration/configuration/TicketIntegrationAutoConfiguration.java
@@ -10,18 +10,18 @@ import de.muenchen.oss.digiwf.process.api.config.api.ProcessConfigApi;
 import de.muenchen.oss.digiwf.s3.integration.client.properties.SupportedFileExtensions;
 import de.muenchen.oss.digiwf.s3.integration.client.repository.DocumentStorageFileRepository;
 import de.muenchen.oss.digiwf.s3.integration.client.repository.DocumentStorageFolderRepository;
-import de.muenchen.oss.digiwf.s3.integration.client.service.S3DomainProvider;
 import de.muenchen.oss.digiwf.s3.integration.client.service.FileExtensionService;
+import de.muenchen.oss.digiwf.s3.integration.client.service.S3DomainProvider;
 import de.muenchen.oss.digiwf.s3.integration.client.service.S3StorageUrlProvider;
-import de.muenchen.oss.digiwf.ticket.integration.adapter.in.streaming.TicketMessageProcessor;
+import de.muenchen.oss.digiwf.ticket.integration.adapter.in.streaming.TicketStreamingAdapter;
 import de.muenchen.oss.digiwf.ticket.integration.adapter.in.streaming.WriteArticleDto;
 import de.muenchen.oss.digiwf.ticket.integration.adapter.out.s3.S3Adapter;
 import de.muenchen.oss.digiwf.ticket.integration.adapter.out.zammad.ZammadAdapter;
-import de.muenchen.oss.digiwf.ticket.integration.adapter.zammad.api.TicketsApi;
-import de.muenchen.oss.digiwf.ticket.integration.application.WriteArticleUseCase;
+import de.muenchen.oss.digiwf.ticket.integration.adapter.out.zammad.api.TicketsApi;
 import de.muenchen.oss.digiwf.ticket.integration.application.port.in.WriteArticleInPort;
 import de.muenchen.oss.digiwf.ticket.integration.application.port.out.LoadFileOutPort;
 import de.muenchen.oss.digiwf.ticket.integration.application.port.out.TicketOutPort;
+import de.muenchen.oss.digiwf.ticket.integration.application.usecase.WriteArticleUseCase;
 import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.context.properties.EnableConfigurationProperties;
@@ -43,8 +43,8 @@ public class TicketIntegrationAutoConfiguration {
 
     @Bean
     public LoadFileOutPort loadFileOutPort(final DocumentStorageFileRepository documentStorageFileRepository,
-            final DocumentStorageFolderRepository documentStorageFolderRepository, final FileExtensionService fileExtensionService,
-            final S3StorageUrlProvider s3StorageUrlProvider) {
+                                           final DocumentStorageFolderRepository documentStorageFolderRepository, final FileExtensionService fileExtensionService,
+                                           final S3StorageUrlProvider s3StorageUrlProvider) {
         return new S3Adapter(documentStorageFileRepository, documentStorageFolderRepository, fileExtensionService, s3StorageUrlProvider);
     }
 
@@ -55,14 +55,14 @@ public class TicketIntegrationAutoConfiguration {
 
     @ConditionalOnMissingBean
     @Bean
-    public TicketMessageProcessor messageProcessor(final WriteArticleInPort writeArticleInPort,
-            final ProcessApi processApi,
-            final ErrorApi errorApi) {
-        return new TicketMessageProcessor(writeArticleInPort, processApi, errorApi);
+    public TicketStreamingAdapter messageProcessor(final WriteArticleInPort writeArticleInPort,
+                                                   final ProcessApi processApi,
+                                                   final ErrorApi errorApi) {
+        return new TicketStreamingAdapter(writeArticleInPort, processApi, errorApi);
     }
 
     @Bean
-    public Consumer<Message<WriteArticleDto>> writeArticle(final TicketMessageProcessor messageProcessor) {
+    public Consumer<Message<WriteArticleDto>> writeArticle(final TicketStreamingAdapter messageProcessor) {
         return messageProcessor.writeArticle();
     }
